+++ src/components/common/modals/secondaryModal/SecondaryModal.tsx
@@ -5,13 +5,10 @@ import ReactModal from 'react-modal';
 import { calcRatio } from '@lib/styles/theme';
 import '../styles.css';
 import CheckButton from '@components/common/buttons/CheckButton';
-import {
-  MODAL_CLOSE_TRANSITION_TIME,
-  MODAL_SLIDE_FROM_POSITION,
-  MODAL_SLIDE_TO_POSITION,
-} from '@lib/constants/MODAL';
-// import useModals from '@lib/hooks/useModals';
+import { MODAL_CLOSE_TRANSITION_TIME } from '@lib/constants/MODAL';
+import useModals from '@lib/hooks/useModals';
 import { modals } from '../Modals';
+import { slideAnimation } from '../slideAnimation';
 
 interface SecondaryModalProps {
   onSubmit: any;
@@ -39,26 +36,19 @@ function SecondaryModal({
   shouldCloseOnOverlayClick = false,
 }: SecondaryModalProps) {
   const [isOpen, setIsOpen] = useState(true);
-  function handleSubmit() {
-    setIsOpen(false);
-    setTimeout(() => {
-      onSubmit();
-    }, MODAL_CLOSE_TRANSITION_TIME);
-  }
-
   let height;
   if (hasBadge) {
     height = '552px';
   } else {
     height = '488px';
   }
-  // const { closeModal } = useModals();
+  const { closeModal } = useModals();
   const reactModalParams = {
     isOpen: isOpen,
     onRequestClose: () => {
       setIsOpen(false);
       setTimeout(() => {
-        // closeModal(modals.secondaryModal);
+        closeModal(modals.secondaryModal);
       }, MODAL_CLOSE_TRANSITION_TIME);
     },
     shouldCloseOnOverlayClick: shouldCloseOnOverlayClick,
@@ -93,6 +83,13 @@ function SecondaryModal({
     },
   };
 
+  const handleSubmit = () => {
+    setIsOpen(false);
+    setTimeout(() => {
+      onSubmit();
+    }, MODAL_CLOSE_TRANSITION_TIME);
+  };
+
   return (
     // @ts-expect-error
     <StyledReactModal {...reactModalParams}>
@@ -119,15 +116,7 @@ function SecondaryModal({
 export default SecondaryModal;
 
 const StyledReactModal = styled(ReactModal)`
-  @keyframes slide {
-    from {
-      transform: translateY(${MODAL_SLIDE_FROM_POSITION});
-    }
-    to {
-      transform: translateY(${MODAL_SLIDE_TO_POSITION});
-    }
-  }
-  animation: slide ${({ theme }) => theme.animation.modalOpen};
+  ${slideAnimation}
 `;
 
 const Content = styled.div`
