+++ src/lib/types/TRiskLevel.ts
@@ -1,0 +1,1 @@
+export type TRiskLevel = '안정' | '보통' | '위험';
